requirements.txt
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.7
cryptography>=42.0
watchdog>=4.0
pytesseract>=0.3.10
Pillow>=10.0
PyPDF2>=3.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0
torch>=2.2
python-dotenv>=1.0
